% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemeColoR.R
\name{schemeColoR}
\alias{schemeColoR}
\title{Scrape color schemes from SchemeColor.com}
\usage{
schemeColoR(palName)
}
\arguments{
\item{palName}{character of length 1 giving the palette name as defined on
SchemeColor.com}
}
\description{
\code{schemeColoR} provides a tool to scrape color palettes
from SchemeColor.com based on the palette name.
}
\details{
SchemeColor.com uses the url format https://www.schemecolor.com/palName.php,
converting spaces in the palette name to dashes. \code{schemeColoR} attempts
to clean-up the given 'palName' removing leading/trailing whitespace and
substituting spaces for dashes.
}
\examples{
schemeColoR("bisexuality-flag-colors")
}
